Agentic RAG/RagAssisstantPhiData/requirements.txt
openai
phidata